Refactor FileHandling
